Updating algorithm to Dr. Li's way
Process:
initialize vertex cover to all true
get the # edges of each vertex and sort the vertices by the number of edges (helps make a more educated guess)
while there isn't a k vertex cover:
go through the list of vertices sorted by size (least to greatest). Try to remove the current one. If it is connected to a vertex that has been removed, you can't remove it. Go to next number. (there may be bugs in there to work out, but that's the general idea)